Add a C source file for qualified file descriptors
One end of error output pipes must be nonblocking.  The listener socket
must be nonblocking and close on exec().  Creation of these file
descriptors is now done in a separate translation unit.  This isolation
makes it easier to use the SOCK_NONBLOCK and SOCK_CLOEXEC constants
which are supported by Linux and POSIX.1-2024 but not by POSIX.1-2017.